Neural Image Aesthetic Score (#45)
* Avatar IPFS

* Fix github workflows

* Fix github workflows

* Filesize limited

* Provisional image upload endpoint

* IPFS Service rename

* IPFS Service rename

* Fix tests

* Review image handling

* IPFS Controller rename

* Fix load script

* Train model

* Model update

* ML API

* Image aesthetic recommender + Image testing in load_data

* Notebook to jupytext

* Fix node dependencies

* Fix node dependencies

* Fix

* Fix

+++ decentraveller-api/web-api/src/orms/image.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, String, Boolean, func, DateTime, Integer
+from sqlalchemy import Column, String, Boolean, func, DateTime, Float
 
 from src.orms import Base
 
@@ -11,5 +11,5 @@ class ImageORM(Base):
 
     hash = Column(String, primary_key=True)
     pinned = Column(Boolean, default=False)
-    score = Column(Integer, default=0)
+    score = Column(Float, default=0.0)
     created_at = Column(DateTime(timezone=True), server_default=func.now())
